clean-up create project flow
