+++ packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/CreateProjectViewAddRole/index.ts
@@ -1,0 +1,1 @@
+export * from "./CreateProjectViewAddRole";
