+++ packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/CreateProjectViewDescription/index.ts
@@ -1,0 +1,1 @@
+export * from "./CreateProjectViewDescription";
