--- packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/CreateProjectViews7/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateProjectViews7";
