+++ packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/index.ts
@@ -1,7 +1,3 @@
-export * from "./CreateProjectViews1";
-export * from "./CreateProjectViews2";
-export * from "./CreateProjectViews3";
-export * from "./CreateProjectViews4";
-export * from "./CreateProjectViews5";
-export * from "./CreateProjectViews6";
-export * from "./CreateProjectViews7";
+export * from "./CreateProjectViewAddRole";
+export * from "./CreateProjectViewDescription";
+export * from "./CreateProjectViewStart";
