+++ packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/older/CreateProjectViews6/CreateProjectViews6.tsx
