+++ packages/ui/src/containers/CreateProjectContainer/CreateProjectViews/older/CreateProjectViews6/index.ts
